Added backend tests for delete tarining category by id
